Move `MemoryShardStore` into `memory` submodule
